server/scripts/requirements.txt
pandas
numpy
scikit-learn
lightgbm
xgboost
psycopg2-binary
python-dotenv
